nit: move implementation of DeletePrefixExpander to memory.rs
